+++ uui-components/src/overlays/Dropdown.tsx
@@ -33,13 +33,13 @@ function DropdownComponent(props: DropdownProps, ref: React.ForwardedRef<HTMLEle
         middleware,
         boundaryElement,
         closeOnEscape = true,
-        pinToToggler = true,
+        fixedBodyPosition = false,
     } = props;
 
     const uuiContext = useContext(UuiContext);
 
     const [uncontrolledOpen, setUncontrolledOpen] = useState(false);
-    const fixedPositionRef = useRef<{ x: number, y: number } | null>(null);
+    const initialBodyPosition = useRef<{ x: number, y: number } | null>(null);
 
     const open = controlledOpen ?? uncontrolledOpen;
     const setOpen = setControlledOpen ?? setUncontrolledOpen;
@@ -437,17 +437,17 @@ function DropdownComponent(props: DropdownProps, ref: React.ForwardedRef<HTMLEle
 
     useEffect(() => {
         if (!open) {
-            fixedPositionRef.current = null;
+            initialBodyPosition.current = null;
         }
     }, [open]);
 
     useEffect(() => {
-        if (!pinToToggler) {
-            if (open && fixedPositionRef.current == null && x && y) {
-                fixedPositionRef.current = { x, y };
+        if (fixedBodyPosition) {
+            if (open && initialBodyPosition.current == null && x && y) {
+                initialBodyPosition.current = { x, y };
             }
         }
-    }, [open, x, y, fixedPositionRef, pinToToggler]);
+    }, [open, x, y, initialBodyPosition, fixedBodyPosition]);
 
     return (
         <>
@@ -463,8 +463,8 @@ function DropdownComponent(props: DropdownProps, ref: React.ForwardedRef<HTMLEle
                             onKeyDown={ floating?.onKeyDown }
                             style={ {
                                 position: strategy,
-                                top: (!pinToToggler && fixedPositionRef.current != null) ? fixedPositionRef.current.y : (y ?? 0),
-                                left: (!pinToToggler && fixedPositionRef.current != null) ? fixedPositionRef.current.x : (x ?? 0),
+                                top: (fixedBodyPosition && initialBodyPosition.current != null) ? initialBodyPosition.current.y : (y ?? 0),
+                                left: (fixedBodyPosition && initialBodyPosition.current != null) ? initialBodyPosition.current.x : (x ?? 0),
                                 zIndex: zIndex != null ? zIndex : layerRef.current?.zIndex,
                             } }
                             data-placement={ finalPlacement }